requirements.txt
apache-airflow-providers-postgres==5.11.3
apache-airflow-providers-google==10.17.0
apache-airflow-providers-smtp==1.8.0

openpyxl
gspread
oauth2client
google-auth
pandas
sqlalchemy
psycopg2-binary
playwright
nest_asyncio
pyperclip
selenium
webdriver-manager
undetected-chromedriver
curl_cffi
python-dotenv